# Conflicts:
#	Applications/BugTracker/pom.xml
#	Applications/ERMailer/pom.xml
#	Applications/JavaMonitor/pom.xml
#	Applications/Uber/pom.xml
#	Applications/pom.xml
#	Applications/wotaskd/pom.xml
#	Examples/Ajax/AjaxExample/pom.xml
#	Examples/Ajax/AjaxExample2/pom.xml
#	Examples/Ajax/pom.xml
#	Examples/D2W/D2WComponentTour/pom.xml
#	Examples/D2W/ERModernMoviesDemo/pom.xml
#	Examples/D2W/ERMovies/pom.xml
#	Examples/D2W/pom.xml
#	Examples/ERJaxWS/WebService_Created/pom.xml
#	Examples/ERJaxWS/WebService_Imported/pom.xml
#	Examples/ERJaxWS/WebService_Stateful/pom.xml
#	Examples/ERJaxWS/pom.xml
#	Examples/ERXPartials/ERXPartialBaseModel/pom.xml
#	Examples/ERXPartials/ERXPartialsExampleApp/pom.xml
#	Examples/ERXPartials/ERXPartialsExampleModel/pom.xml
#	Examples/ERXPartials/pom.xml
#	Examples/Misc/ERComponentTour/pom.xml
#	Examples/Misc/ERIndexingExample/pom.xml
#	Examples/Misc/ERRestRouteExample/pom.xml
#	Examples/Misc/ERSeleniumExample/pom.xml
#	Examples/Misc/TickTockMan/pom.xml
#	Examples/Misc/pom.xml
#	Examples/Reporting/pom.xml
#	Examples/pom.xml
#	Frameworks/Ajax/Ajax/pom.xml
#	Frameworks/Ajax/ERCoolComponents/pom.xml
#	Frameworks/Ajax/ERJQuery/pom.xml
#	Frameworks/Ajax/ERJQueryMobile/pom.xml
#	Frameworks/Ajax/ERPrototaculous/pom.xml
#	Frameworks/Ajax/GoogleChart/pom.xml
#	Frameworks/Ajax/MooTools/pom.xml
#	Frameworks/Ajax/pom.xml
#	Frameworks/BusinessLogic/BTBusinessLogic/pom.xml
#	Frameworks/BusinessLogic/ERAttachment/pom.xml
#	Frameworks/BusinessLogic/ERCoreBusinessLogic/pom.xml
#	Frameworks/BusinessLogic/ERMoviesLogic/pom.xml
#	Frameworks/BusinessLogic/ERTaggable/pom.xml
#	Frameworks/BusinessLogic/WOPayPal/pom.xml
#	Frameworks/BusinessLogic/pom.xml
#	Frameworks/Core/ERDirectToWeb/pom.xml
#	Frameworks/Core/ERExtensions/pom.xml
#	Frameworks/Core/ERJavaMail/pom.xml
#	Frameworks/Core/ERNeutralLook/pom.xml
#	Frameworks/Core/ERPrototypes/pom.xml
#	Frameworks/Core/JavaWOExtensions/pom.xml
#	Frameworks/Core/WOOgnl/pom.xml
#	Frameworks/Core/pom.xml
#	Frameworks/D2W/ERModernDefaultSkin/pom.xml
#	Frameworks/D2W/ERModernDirectToWeb/pom.xml
#	Frameworks/D2W/ERModernLook/pom.xml
#	Frameworks/D2W/ERMootoolsDirectToWeb/pom.xml
#	Frameworks/D2W/pom.xml
#	Frameworks/EOAdaptors/JavaERJDBCAdaptor/pom.xml
#	Frameworks/EOAdaptors/JavaFSAdaptor/pom.xml
#	Frameworks/EOAdaptors/JavaLuceneAdaptor/pom.xml
#	Frameworks/EOAdaptors/JavaMemoryAdaptor/pom.xml
#	Frameworks/EOAdaptors/JavaNeo4JAdaptor/pom.xml
#	Frameworks/EOAdaptors/pom.xml
#	Frameworks/EOF/ERAttributeExtension/pom.xml
#	Frameworks/EOF/ERChangeNotificationJMS/pom.xml
#	Frameworks/EOF/ERIndexing/pom.xml
#	Frameworks/EOF/ERJGroupsSynchronizer/pom.xml
#	Frameworks/EOF/ERRest/pom.xml
#	Frameworks/EOF/pom.xml
#	Frameworks/Excel/ERExcelLook/pom.xml
#	Frameworks/Excel/ExcelGenerator/pom.xml
#	Frameworks/Excel/pom.xml
#	Frameworks/Misc/ERCalendar/pom.xml
#	Frameworks/Misc/ERCaptcha/pom.xml
#	Frameworks/Misc/ERCayenne/pom.xml
#	Frameworks/Misc/ERChronic/pom.xml
#	Frameworks/Misc/ERDistribution/pom.xml
#	Frameworks/Misc/ERFirebug/pom.xml
#	Frameworks/Misc/ERJaxWS/pom.xml
#	Frameworks/Misc/ERPDFGeneration/pom.xml
#	Frameworks/Misc/ERPersistentSessionStorage/pom.xml
#	Frameworks/Misc/ERProfiling/pom.xml
#	Frameworks/Misc/ERQuartzScheduler/pom.xml
#	Frameworks/Misc/ERSelenium/pom.xml
#	Frameworks/Misc/ERXiss/pom.xml
#	Frameworks/Misc/JavaMonitorFramework/pom.xml
#	Frameworks/Misc/SnapshotExplorer/pom.xml
#	Frameworks/Misc/WOJRebel/pom.xml
#	Frameworks/Misc/WOLips/pom.xml
#	Frameworks/Misc/pom.xml
#	Frameworks/PlugIns/DB2PlugIn/pom.xml
#	Frameworks/PlugIns/DerbyPlugIn/pom.xml
#	Frameworks/PlugIns/EROraclePlugIn/pom.xml
#	Frameworks/PlugIns/FrontBasePlugIn/pom.xml
#	Frameworks/PlugIns/H2PlugIn/pom.xml
#	Frameworks/PlugIns/MySQLPlugIn/pom.xml
#	Frameworks/PlugIns/PostgresqlPlugIn/pom.xml
#	Frameworks/PlugIns/pom.xml
#	Frameworks/Reporting/DRGrouping/pom.xml
#	Frameworks/Reporting/ERJasperReports/pom.xml
#	Frameworks/Reporting/ERPlot/pom.xml
#	Frameworks/Reporting/WRReporting/pom.xml
#	Frameworks/Reporting/pom.xml
#	Frameworks/WOAdaptors/ERWOAdaptor/pom.xml
#	Frameworks/WOAdaptors/pom.xml
#	Frameworks/pom.xml
#	Utilities/WOInstall/pom.xml
#	Utilities/pom.xml
#	pom.xml

+++ Utilities/Adaptors/Adaptor/appcfg.c
@@ -1232,7 +1232,7 @@ static char *_retrieveServerInfo(ConfigServer *server, net_fd s, int *len, char 
    }
    tr_close(c, AC_INVALID_HANDLE, 0);
    if (resp != NULL) {
-      *len = resp->content_length;
+      *len = (int) resp->content_length;
       *content_type = '\0';
       if (resp->status == HTTP_OK) {
          /* may be NOT_MODIFIED (or worse!) */
