+++ Utilities/Adaptors/Adaptor/request.h
@@ -51,7 +51,7 @@ typedef struct _HTTPRequest {
 	char *request_str;			/* the http request (includes CRLF) */
 	void *headers;				/* (strtbl *) but you don't need to know */
 	void *api_handle;			/* api specific pointer */
-	unsigned content_length;
+	unsigned long content_length;
 	void *content;
         unsigned content_buffer_size;
         req_getMoreContentCallback getMoreContent;
